sign in mechanics up

+++ src/employee/Employee.cpp
@@ -1,5 +1,5 @@
 #include "Employee.h"
 
 Employee::Employee(
-    const std::string &workID, const uint16_t &social, const AccessLevel &accessLevel)
+    const std::string &workID, const uint32_t &social, const AccessLevel &accessLevel)
     : workID_(workID), social_(social), accessLevel_(accessLevel) {}
